utils/pipeline/requirements-dev.txt
# Development dependencies with specific versions to avoid compatibility issues
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
pytest-xdist==3.3.1
pytest-sugar==0.9.7

# Type checking
mypy==1.5.1
types-PyYAML==6.0.12.12

# Linting and formatting
ruff==0.0.292
black==23.7.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Core dependencies
pyyaml>=6.0.1
typing-extensions>=4.7.1
